requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
pip==24.2
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==75.1.0
shortuuid==1.0.13
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.44.0
